Fixed nullptr error by removing the references that i made to non pointer objects and re-adding poniter to them. moved Sprites and UI that are created in the GameManager to the header file so they can be deleted and removed from memory. changed foreach loops to be const auto& instead of direct class definition, to have the code more flexible if change to a different list type.

+++ TowerDefence/TowerDefence/Collider.cpp
@@ -32,7 +32,7 @@ void Collider::DrawDebugCircle(SDL_Renderer* renderer)
 {
 	if (radius > 0)
 	{
-		for (size_t i = 0; i < 360; i += debugCircleQuality)
+		for (unsigned int i = 0; i < 360; i += debugCircleQuality)
 		{
 			debugCirclePosition = Vector2D(position.x + radius * std::cos(i), position.y + radius * std::sin(i));
 			SDL_SetRenderDrawColor(renderer, 230, 0, 126, 255);
